refactor(Validation): 클래스 디렉터리 변경
domain 디렉터리로 이동
